:art: Reorg view tests
